feat: add mocked VehiclesService
Add a mocked (in-memory only) implementation of a VehiclesService so that we can start scaffolding out more of the application.

Also adds Jest, and a first test to validate that creating an entity works.

This test is valuable right now to verify that we are able to connect to & use whatever type of persistence we are planning on (in memory for now).

Later on, it would be implicitly verified by things like e2e tests, so it would be something we either further develop in another direction or perhaps grow out of.