chore: add dev-oriented data load
Add a very crude 'data-load' to the application, so that we are at least guaranteed some data on start up.

Note that it checks for existing seed data before inserting, for when it is eventually run against a live DB.

This is a minor accomodation only, and there could be a lot more done to harden this/even cover some common scenarios (updated seed collection, for instance).

For now these scenarios must be handled manually by the user.